Modify ProfileSettingVC and add PrivacyPolicy enum file
